W2-G3: builtin float at two numpy roots; legacy schedule spans 24 h (#283, #284) (#444)
* tests: failing checks for numpy roots (#283) and 24 h schedule span (#284)

Import the production symbols. On this tree they fail: compute_solar_gain
and _deferred_energy_cost still return np.float64, _build_data_dict still
publishes 5h45m, and the two state strings stay plural at 1.

* fix: return builtin float from compute_solar_gain and _deferred_energy_cost (#283)

Two roots, not a _plain_types wrap of _build_data_dict. float() at the
solar-gain product collapses the trajectory and the daylight twin; float()
at the deferred-cost product collapses climate predicted_savings, which
inherits numpy through the subtraction.

* fix: publish the full 24 h on the two legacy schedule sensors (#284)

Drop the [:24]/[1:25] slices in _build_data_dict so schedule and
dhw_schedule span the same horizon as space_plan.forecast. Singular
grammar at 1; README DHW row now matches the 24 h claim.

+++ custom_components/heatpump_optimizer/optimizer.py
@@ -5344,7 +5344,7 @@ class HeatPumpOptimizer:
                     - _capped(optimized_end.dhw_temperature, "dhw")
                 )
                 electrical += dhw_gap / cop_dhw - dhw_gap / cop
-        return electrical * refill_price
+        return float(electrical * refill_price)
 
     def _compute_baseline_power(
         self,
